+++ packages/platforms/browser/tests/auto-instrumentation/network-request-plugin.test.ts
@@ -1,11 +1,16 @@
-import { spanContextEquals } from '@bugsnag/core-performance'
-import { MockSpanFactory, createConfiguration, createTestClient } from '@bugsnag/js-performance-test-utilities'
+/**
+ * @jest-environment jsdom
+ */
+
+import { DefaultSpanContextStorage, type SpanContextStorage, spanContextEquals } from '@bugsnag/core-performance'
+import { ControllableBackgroundingListener, MockSpanFactory, createConfiguration, createTestClient } from '@bugsnag/js-performance-test-utilities'
 import { type BrowserSchema, type BrowserConfiguration } from '../../lib/config'
 import { RequestTracker, type RequestStartCallback } from '@bugsnag/request-tracker-performance'
 import { NetworkRequestPlugin } from '../../lib/auto-instrumentation/network-request-plugin'
 
 const ENDPOINT = 'http://traces.endpoint'
 const TEST_URL = 'http://test-url.com/'
+const SAME_ORIGIN_TEST_URL = 'http://localhost/my-api'
 
 class MockRequestTracker extends RequestTracker {
   onStart = jest.fn((startCallback: RequestStartCallback) => {
@@ -17,51 +22,64 @@ describe('network span plugin', () => {
   let xhrTracker: MockRequestTracker
   let fetchTracker: MockRequestTracker
   let spanFactory: MockSpanFactory<BrowserConfiguration>
+  let spanContextStorage: SpanContextStorage
 
   beforeEach(() => {
     xhrTracker = new MockRequestTracker()
     fetchTracker = new MockRequestTracker()
     spanFactory = new MockSpanFactory()
+
+    spanContextStorage = new DefaultSpanContextStorage(new ControllableBackgroundingListener())
   })
 
   it('tracks requests when autoInstrumentNetworkRequests = true', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
-    expect(xhrTracker.onStart).not.toHaveBeenCalled()
-    expect(fetchTracker.onStart).not.toHaveBeenCalled()
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
       autoInstrumentNetworkRequests: true
     }))
 
-    expect(xhrTracker.onStart).toHaveBeenCalled()
-    expect(fetchTracker.onStart).toHaveBeenCalled()
+    const res = fetchTracker.start({ type: 'fetch', method: 'GET', url: SAME_ORIGIN_TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([])
+
+    const res2 = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 2 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(res2.extraRequestHeaders).toEqual([])
   })
 
   it('starts a span on request start', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
       autoInstrumentNetworkRequests: true
     }))
 
-    fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
+    const res1 = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
     expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(res1.extraRequestHeaders).toEqual([])
 
-    xhrTracker.start({ type: 'xmlhttprequest', method: 'POST', url: TEST_URL, startTime: 2 })
+    const res2 = xhrTracker.start({ type: 'xmlhttprequest', method: 'POST', url: TEST_URL, startTime: 2 })
     expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/POST', { startTime: 2, makeCurrentContext: false })
+    expect(res2.extraRequestHeaders).toEqual([])
+
+    // traceparent headers are not added to same-origin requests by default
+    const res3 = xhrTracker.start({ type: 'xmlhttprequest', method: 'POST', url: SAME_ORIGIN_TEST_URL, startTime: 2 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/POST', { startTime: 2, makeCurrentContext: false })
+    expect(res3.extraRequestHeaders).toEqual([])
   })
 
   it('ends a span on request end', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
       autoInstrumentNetworkRequests: true
     }))
 
-    const endRequest = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
+    const { onRequestEnd: endRequest } = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
     expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
     expect(spanFactory.endSpan).not.toHaveBeenCalled()
 
@@ -80,7 +98,7 @@ describe('network span plugin', () => {
   })
 
   it('does not track requests when autoInstrumentNetworkRequests = false', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
@@ -92,7 +110,7 @@ describe('network span plugin', () => {
   })
 
   it('does not track requests to the configured traces endpoint', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
@@ -113,7 +131,7 @@ describe('network span plugin', () => {
   ]
 
   it.each(expectedProtocols)('tracks requests over all expected protocols', (url) => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
@@ -137,7 +155,7 @@ describe('network span plugin', () => {
   ]
 
   it.each(unexpectedProtocols)('does not track requests over unexpected protocols', (url) => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
@@ -149,14 +167,14 @@ describe('network span plugin', () => {
   })
 
   it('discards the span if the status is 0', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
       autoInstrumentNetworkRequests: true
     }))
 
-    const endRequest = xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: TEST_URL, startTime: 1 })
+    const { onRequestEnd: endRequest } = xhrTracker.start({ type: 'xmlhttprequest', method: 'GET', url: TEST_URL, startTime: 1 })
     expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
 
     endRequest({ endTime: 2, state: 'error' })
@@ -164,14 +182,14 @@ describe('network span plugin', () => {
   })
 
   it('discards the span if there is an error', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
 
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
       autoInstrumentNetworkRequests: true
     }))
 
-    const endRequest = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
+    const { onRequestEnd: endRequest } = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
     expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
 
     endRequest({ state: 'error', error: new Error('woopsy'), endTime: 2 })
@@ -179,7 +197,11 @@ describe('network span plugin', () => {
   })
 
   it('does not push network spans to the context stack', () => {
-    const client = createTestClient<BrowserSchema, BrowserConfiguration>({ plugins: (spanFactory) => [new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)] })
+    const client = createTestClient<BrowserSchema, BrowserConfiguration>({
+      plugins: (spanFactory, spanContextStorage) => [
+        new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
+      ]
+    })
     const rootSpan = client.startSpan('root span')
 
     fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
@@ -189,7 +211,7 @@ describe('network span plugin', () => {
   })
 
   it('prevents creating a span when networkRequestCallback returns null', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
       autoInstrumentNetworkRequests: true,
@@ -203,8 +225,59 @@ describe('network span plugin', () => {
     expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
   })
 
+  it('prevents creating a span when networkRequestCallback returns { url: null }', () => {
+    // if { url: null } is returned from the networkRequestCallback, a span is not created
+    // and trace propagation headers are not added if there is no span context as well
+    spanContextStorage.push({
+      id: 'abc123',
+      traceId: 'xyz456',
+      samplingRate: 0.1,
+      isValid: () => true
+    })
+
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
+    plugin.configure(createConfiguration<BrowserConfiguration>({
+      endpoint: ENDPOINT,
+      autoInstrumentNetworkRequests: true,
+      networkRequestCallback: (networkRequestInfo) => ({ ...networkRequestInfo, url: null })
+    }))
+
+    const res = fetchTracker.start({ type: 'fetch', method: 'GET', url: SAME_ORIGIN_TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).not.toHaveBeenCalled()
+    expect(res.extraRequestHeaders).toEqual([])
+  })
+
+  it('prevents adding trace propagation headers when networkRequestCallback returns { propagateTraceContext: false }', () => {
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
+    plugin.configure(createConfiguration<BrowserConfiguration>({
+      endpoint: ENDPOINT,
+      autoInstrumentNetworkRequests: true,
+      networkRequestCallback: (networkRequestInfo) => ({ ...networkRequestInfo, propagateTraceContext: false })
+    }))
+
+    const res = fetchTracker.start({ type: 'fetch', method: 'GET', url: SAME_ORIGIN_TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([])
+  })
+
+  it('adds trace propagation headers when networkRequestCallback returns { propagateTraceContext: true }', () => {
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
+    plugin.configure(createConfiguration<BrowserConfiguration>({
+      endpoint: ENDPOINT,
+      autoInstrumentNetworkRequests: true,
+      networkRequestCallback: (networkRequestInfo) => ({ ...networkRequestInfo, propagateTraceContext: true })
+    }))
+
+    // by default when the origin is different trace propagation headers are not added, unless the networkRequestCallback returns { propagateTraceContext: true }
+    const res = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
+    expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
+    expect(res.extraRequestHeaders).toEqual([
+      { traceparent: '00-a random 128 bit string-a random 64 bit string-01' }
+    ])
+  })
+
   it('uses a modified url from networkRequestCallback', () => {
-    const plugin = new NetworkRequestPlugin(spanFactory, fetchTracker, xhrTracker)
+    const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
     plugin.configure(createConfiguration<BrowserConfiguration>({
       endpoint: ENDPOINT,
       autoInstrumentNetworkRequests: true,
@@ -214,7 +287,7 @@ describe('network span plugin', () => {
       })
     }))
 
-    const endRequest = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
+    const { onRequestEnd: endRequest } = fetchTracker.start({ type: 'fetch', method: 'GET', url: TEST_URL, startTime: 1 })
     expect(spanFactory.startSpan).toHaveBeenCalledWith('[HTTP]/GET', { startTime: 1, makeCurrentContext: false })
 
     endRequest({ status: 200, endTime: 2, state: 'success' })
@@ -229,5 +302,91 @@ describe('network span plugin', () => {
     expect(span).toHaveAttribute('http.url', 'modified-url')
     expect(span).toHaveAttribute('http.method', 'GET')
     expect(span).toHaveAttribute('http.status_code', 200)
+  })
+
+  describe('generating traceparent extraRequestHeaders', () => {
+    describe('when a network span is being created for the request', () => {
+      it('generates a traceparent extraRequestHeader if propagateTraceContext is set to true', () => {
+        spanContextStorage.push({
+          id: 'abc123',
+          traceId: 'xyz456',
+          samplingRate: 0.1,
+          isValid: () => true
+        })
+        const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
+
+        plugin.configure(createConfiguration<BrowserConfiguration>({
+          endpoint: ENDPOINT,
+          autoInstrumentNetworkRequests: true,
+          networkRequestCallback: (requestInfo) => {
+            requestInfo.propagateTraceContext = true
+            return requestInfo
+          }
+        }))
+
+        const res = fetchTracker.start({ type: 'fetch', method: 'GET', url: 'https://my-api.com/users', startTime: 1 })
+
+        expect(res.extraRequestHeaders).toEqual([
+          { traceparent: '00-a random 128 bit string-a random 64 bit string-01' }
+        ])
+
+        const res2 = fetchTracker.start({ type: 'fetch', method: 'GET', url: '/users', startTime: 1 })
+
+        expect(res2.extraRequestHeaders).toEqual([
+          { traceparent: '00-a random 128 bit string-a random 64 bit string-01' }
+        ])
+      })
+
+      it('does not generate a traceparent extraRequestHeader if propagateTraceContext is set to false', () => {
+        spanContextStorage.push({
+          id: 'abc123',
+          traceId: 'xyz456',
+          samplingRate: 0.1,
+          isValid: () => true
+        })
+        const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
+
+        plugin.configure(createConfiguration<BrowserConfiguration>({
+          endpoint: ENDPOINT,
+          autoInstrumentNetworkRequests: true,
+          networkRequestCallback: (requestInfo) => {
+            requestInfo.propagateTraceContext = false
+            return requestInfo
+          }
+        }))
+
+        const res = fetchTracker.start({ type: 'fetch', method: 'GET', url: 'https://not-my-api.com/users', startTime: 1 })
+
+        expect(res.extraRequestHeaders).toEqual([])
+      })
+    })
+
+    describe('when a network span is not being created for the request', () => {
+      it('uses the span context', () => {
+        spanContextStorage.push({
+          id: 'abc123',
+          traceId: 'xyz456',
+          samplingRate: 0.1,
+          isValid: () => true
+        })
+        const plugin = new NetworkRequestPlugin(spanFactory, spanContextStorage, fetchTracker, xhrTracker)
+
+        plugin.configure(createConfiguration<BrowserConfiguration>({
+          endpoint: ENDPOINT,
+          autoInstrumentNetworkRequests: true,
+          networkRequestCallback: (requestInfo) => {
+            requestInfo.propagateTraceContext = true
+            requestInfo.url = null
+            return requestInfo
+          }
+        }))
+
+        const res = fetchTracker.start({ type: 'fetch', method: 'GET', url: 'https://my-api.com/users', startTime: 1 })
+
+        expect(res.extraRequestHeaders).toEqual([
+          { traceparent: '00-xyz456-abc123-01' }
+        ])
+      })
+    })
   })
 })