+++ packages/plugin-react-native-navigation/lib/index.ts
@@ -1,0 +1,2 @@
+export { default } from './react-native-navigation-plugin'
+export { CompleteNavigation } from './CompleteNavigation'