+++ packages/request-tracker/lib/request-tracker.ts
@@ -19,7 +19,7 @@ export interface RequestEndContextError {
 
 export type RequestEndContext = RequestEndContextSuccess | RequestEndContextError
 
-export type RequestStartCallback = (context: RequestStartContext) => RequestEndCallback | undefined
+export type RequestStartCallback = (context: RequestStartContext) => { onRequestEnd?: RequestEndCallback, extraRequestHeaders?: Record<string, string> } | undefined
 
 export type RequestEndCallback = (context: RequestEndContext) => void
 
@@ -31,16 +31,30 @@ export class RequestTracker {
   }
 
   start (context: RequestStartContext) {
-    const endCallbacks: RequestEndCallback[] = []
+    const results: Array<ReturnType<RequestStartCallback>> = []
     for (const startCallback of this.callbacks) {
-      const endCallback = startCallback(context)
-      if (endCallback) endCallbacks.push(endCallback)
+      const result = startCallback(context)
+      if (result) results.push(result)
     }
 
-    return (endContext: RequestEndContext) => {
-      for (const endCallback of endCallbacks) {
-        endCallback(endContext)
-      }
+    return {
+      onRequestEnd: (endContext: RequestEndContext) => {
+        for (const result of results) {
+          if (result && result.onRequestEnd) {
+            result.onRequestEnd(endContext)
+          }
+        }
+      },
+      extraRequestHeaders: results.map((result) => {
+        if (result && result.extraRequestHeaders) {
+          return result.extraRequestHeaders
+        }
+        return undefined
+      }).filter(isDefined)
     }
   }
+}
+
+function isDefined<T> (argument: T | undefined): argument is T {
+  return argument !== undefined
 }