+++ test/browser/features/fixtures/packages/traceparent/src/index.js
@@ -1,0 +1,30 @@
+import BugsnagPerformance from '@bugsnag/browser-performance'
+
+const parameters = new URLSearchParams(window.location.search)
+const apiKey = parameters.get('api_key')
+const endpoint = parameters.get('endpoint')
+const reflectEndpoint = '/reflect?status=200&delay_ms=0'
+
+BugsnagPerformance.start({
+  apiKey, 
+  endpoint,
+  autoInstrumentFullPageLoads: false,
+  autoInstrumentNetworkRequests: true,
+  networkRequestCallback: (requestInfo) => {
+    requestInfo.propagateTraceContext = true
+    return requestInfo;
+  },
+  maximumBatchSize: 1,
+ })
+
+document.getElementById("xhr").addEventListener("click", () => {
+  const xhr = new XMLHttpRequest()
+  xhr.open('GET', reflectEndpoint)
+  xhr.setRequestHeader('X-Test-Header', 'test')
+  xhr.send()
+})
+
+document.getElementById("fetch").addEventListener("click", () => {
+  fetch(reflectEndpoint, { headers: { 'X-Test-Header': 'test' }})
+})
+